+++ micro-judge/src/io/messages.rs
@@ -1,6 +1,6 @@
 use super::topics::*;
 use crate::model::*;
-use log::{error, warn};
+use log::error;
 use redis::streams::StreamReadOptions;
 use redis::{Client, Commands};
 use redis_streams::*;
@@ -44,27 +44,6 @@ pub fn read_sorted(
     }
 
     Ok(answer)
-}
-
-const GROUP_NAME: &str = "micro-judge";
-pub fn create_consumer_groups(topics: &StreamTopics, client: &Client) -> String {
-    let mut conn = client.get_connection().expect("group create conn");
-    let mm: Result<(), _> = conn.xgroup_create(&topics.make_move_cmd, GROUP_NAME, "$");
-    if let Err(e) = mm {
-        warn!(
-            "Ignoring error creating MakeMoveCmd consumer group (it probably exists already) {:?}",
-            e
-        );
-    }
-    let gs: Result<(), _> = conn.xgroup_create(&topics.game_states_changelog, GROUP_NAME, "$");
-    if let Err(e) = gs {
-        warn!(
-            "Ignoring error creating GameStates consumer group (it probably exists already) {:?}",
-            e
-        );
-    }
-
-    GROUP_NAME.to_string()
 }
 
 pub fn ack(
